fix: ajustar tipos das migrations Postgres (BIGINT, BOOLEAN, DOUBLE PRECISION) e desativar schema-validation

+++ src/main/resources/db/migration/postgresql/V2__criar_tabela_usuario.sql
@@ -1,7 +1,7 @@
 CREATE SEQUENCE seq_tbl_usuarios START WITH 1 INCREMENT BY 1;
 
 CREATE TABLE tbl_usuarios (
-    usuario_id INTEGER NOT NULL PRIMARY KEY,
+    usuario_id BIGINT NOT NULL PRIMARY KEY,
     nome       VARCHAR(100) NOT NULL,
     email      VARCHAR(100) UNIQUE NOT NULL,
     senha      VARCHAR(20) NOT NULL,
